An alternate Technic Solder installation
These are the instructions that worked for me. They continue to work. The reason is they eliminate choice in the same way that a Utopian governmental system would or a Single Pay healthcare system might do. You get no choice in what to do. Deviation will result in you having many headaches. You don't want that. The STATE IS ALL.

This comment not really tongue-in-cheek. I personally don't like Utopian concepts, socialism of any kind, and single payer health care systems. The only people we should be taking care of are those unable to physically work of whom have no family or friends willing to support them. You see that includes the disabled and infirm too. People should be evaluate to see if they are lying about their status and if found to be doing so asked which country they will like to be deported too. Because if you are going to Defraud your country then you don't want to be a citizen of it and should live elsewhere giving up all of your rights to your original country in the process. This a full exile and I think very humane.

I work for myself and I make enough coupled with chicken ranching and gardening to get by comfortably, own several computers, have 6 pet dogs, own and drive two cars, etc. I don't live like an extravagant person. I keep my wants way down below my needs. I survive comfortably on $1500/month (2017 A.D.)